examples/demos/django-drf/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
